Reference

ArviZPlots.distplot

ArviZPlots.distplot!

ArviZPlots.ecdfplot

ArviZPlots.kde2dplot

ArviZPlots.kde2dplot!

ArviZPlots.kdeplot

ArviZPlots.kdeplot!

ArviZPlots.rugplot

RecipesBase.plot

RecipesBase.plot

API

ArviZPlots.distplot! — Method

distplot(x; kwargs...)
distplot!(x; kwargs...)

Plot distribution as histogram, KDE, or ECDF.

By default continuous variables are plotted using KDEs and discrete ones using histograms

Positional Arguments

x: values to plot

Keyword Arguments

kind=:auto: By default (:auto), continuous variables are plotted using KDEs and discrete ones using histograms. The following values can override this behavior:

kind=:kde: Plot using a kernel density estimate (see kdeplot)

kind=:hist: Plot using a histogram (see [Plots.histogram]

kind=:ecdf: Plot using an empirical cumulative distribution function (see ecdfplot)

kwargs: Additional attributes understood by Plots.jl. In particular, see keyword arguments for kdeplot and histogram.

ArviZPlots.distplot — Method

distplot(x; kwargs...)
distplot!(x; kwargs...)

Plot distribution as histogram, KDE, or ECDF.

By default continuous variables are plotted using KDEs and discrete ones using histograms

Positional Arguments

x: values to plot

Keyword Arguments

kind=:auto: By default (:auto), continuous variables are plotted using KDEs and discrete ones using histograms. The following values can override this behavior:

kind=:kde: Plot using a kernel density estimate (see kdeplot)

kind=:hist: Plot using a histogram (see [Plots.histogram]

kind=:ecdf: Plot using an empirical cumulative distribution function (see ecdfplot)

kwargs: Additional attributes understood by Plots.jl. In particular, see keyword arguments for kdeplot and histogram.

ArviZPlots.kde2dplot! — Method

kde2dplot(x, y; kwargs...)
kde2dplot!(x, y; kwargs...)

Plot a 2D kernel density estimate (KDE) taking into account boundary conditions.

Positional Arguments

x: x coordinates of the values

y: y coordinates of the values

Keyword Arguments

contour=true: If true, plot the 2D KDE using contours, otherwise plot a heatmap.

kwargs: Additional attributes understood by Plots.jl.

ArviZPlots.kde2dplot — Method

kde2dplot(x, y; kwargs...)
kde2dplot!(x, y; kwargs...)

Plot a 2D kernel density estimate (KDE) taking into account boundary conditions.

Positional Arguments

x: x coordinates of the values

y: y coordinates of the values

Keyword Arguments

contour=true: If true, plot the 2D KDE using contours, otherwise plot a heatmap.

kwargs: Additional attributes understood by Plots.jl.

ArviZPlots.kdeplot! — Method

kdeplot(x[, y]; kwargs...)
kdeplot!(x[, y]; kwargs...)

Plot a 1D or 2D kernel density estimate (KDE) taking into account boundary conditions.

Positional Arguments

x: Values to plot.

y: Values to plot. If present, a 2D KDE will be estimated. See kde2dplot.

Keyword Arguments

cumulative=false: If true, plot the estimated cumulative distribution function

showrug=false: If true, adds a rugplot

rugspace=0.05: control the y position of the rugplot. The larger this number the lower the rugplot

kwargs: Additional attributes understood by Plots.jl. In particular, see keyword arguments for kde2dplot and rugplot.

ArviZPlots.kdeplot — Method

kdeplot(x[, y]; kwargs...)
kdeplot!(x[, y]; kwargs...)

Plot a 1D or 2D kernel density estimate (KDE) taking into account boundary conditions.

Positional Arguments

x: Values to plot.

y: Values to plot. If present, a 2D KDE will be estimated. See kde2dplot.

Keyword Arguments

cumulative=false: If true, plot the estimated cumulative distribution function

showrug=false: If true, adds a rugplot

rugspace=0.05: control the y position of the rugplot. The larger this number the lower the rugplot

kwargs: Additional attributes understood by Plots.jl. In particular, see keyword arguments for kde2dplot and rugplot.

Miscellaneous

ArviZPlots.ecdfplot — Method

ecdfplot(x; kwargs...)
ecdfplot!(x; kwargs...)

Plot the empirical cumulative distribution function (ECDF) of the data x.

Positional Arguments

x: Values to plot

Keyword Arguments

weights=nothing: Vector of weights to be used for a weighted ECDF

kwargs: Additional attributes understood by Plots.jl

ArviZPlots.rugplot — Method

rugplot(x; kwargs...)
rugplot!(x; kwargs...)

Make a rugplot.

Positional Arguments

x: values

Keyword Arguments

rugmarkershape=:auto: shape of marker, defaulting to perpendicular to the data axis

rugsize=8: size (height) of marker

rugcolor=:auto: color of marker

rugalpha=:auto: alpha value (transparency) of marker

rugposition=0: position of rug on non-data axis

kwargs: Additional attributes understood by Plots.jl

RecipesBase.plot — Method

plot(data::ArviZ.Dataset; kwargs...)
plot!(data::ArviZ.Dataset; kwargs...)

Plot the specified group of data using kwargs attributes recognized by Plots.

data is split into individual subplots based on the provided keyword arguments. Chain number for each draw is stored in the plot attribute chain and can be used by series recipes.

Positional Arguments

data: an ArviZ.Dataset object.

Keyword Arguments

var_names=nothing: Variables to be plotted. Prefix the variables by ~ to exclude them from the plot. Behavior modified by filter_vars.

filter_vars=nothing: interpret var_names as

filter_vars=nothing: real variables names

filter_vars="like": substrings of the real variables names

filter_vars="regex": regular expressions on the real variables names

coords=Dict(): Specific coordinates of var_names to be plotted

kwargs: Additional attributes understood by Plots.jl

RecipesBase.plot — Method

plot(data::InferenceData; groupname=:posterior, kwargs...)
plot!(data::InferenceData; groupname=:posterior, kwargs...)

Plot the specified group of data using kwargs attributes recognized by Plots.